% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_get.R
\name{marawi_get_armm}
\alias{marawi_get_armm}
\alias{marawi_get_armm_all}
\title{Get ARMM Barangay level data masterlist from the Open Marawi Google Drive
Database}
\usage{
marawi_get_armm(
  dataset = c("metadata", "demographics", "bgyfacilities", "conflict")
)

marawi_get_armm_all(tabular = TRUE)
}
\arguments{
\item{dataset}{A string value from either \emph{"metadata"}, \emph{"demographics"},
\emph{"bgyfacilities"}, or \emph{"conflict"} specifying the masterlist data to
retrieve. If not specified, will default to \emph{"metadata"}.}

\item{tabular}{Logical. Should output be flattened into a single data.frame?
Default is TRUE. If FALSE, output is a list of data.frames for each
sheet in the masterlist Excel file.}
}
\value{
A tibble or a list of tibbles of the specified masterlist dataset
available in the ARMM directory of the Open Marawi Database.
}
\description{
Get ARMM Barangay level data masterlist from the Open Marawi Google Drive
Database
}
\examples{
## Retrieve the metadata dataset of the ARMM directory in the Open Marawi
## database
marawi_get_armm(dataset = "metadata")
\dontrun{marawi_get_armm_all()}

}
\author{
Ernest Guevarra
}
