% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_get.R
\name{marawi_get_lanao}
\alias{marawi_get_lanao}
\alias{marawi_get_lanao_all}
\title{Get Lanao del Sur Barangay level data masterlist from the Open Marawi Google
Drive Database}
\usage{
marawi_get_lanao(
  dataset = c("metadata", "demographics", "bgyfacilities", "estab", "modetranspo",
    "conflict", "privateschools", "health")
)

marawi_get_lanao_all(tabular = TRUE)
}
\arguments{
\item{dataset}{A string value from either \emph{"metadata"}, \emph{"demographics"},
\emph{"facilities"}, \emph{"establishment"}, \emph{"mode_transport"}, \emph{"conflict"},
\emph{"private_schools"}, or \emph{"health"} specifying the masterlist data to
retrieve. If not specified, will default to \emph{"metadata"}.}

\item{tabular}{Logical. Should output be flattened into a single data.frame?
Default is TRUE. If FALSE, output is a list of data.frames for each
sheet in the masterlist Excel file.}
}
\value{
A tibble or a list of tibbles of the specified masterlist dataset
available in the Lanao del Sur directory of the Open Marawi Database.
}
\description{
Get Lanao del Sur Barangay level data masterlist from the Open Marawi Google
Drive Database
}
\examples{
## Retrieve the metadata dataset of the Lanao del Sur directory in the Open
## Marawi database
marawi_get_lanao(dataset = "metadata")
marawi_get_lanao_all()

}
\author{
Ernest Guevarra
}
