% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_get.R
\name{marawi_get_marawi}
\alias{marawi_get_marawi}
\alias{marawi_get_marawi_all}
\title{Get Marawi Barangay level data masterlist from the Open Marawi Google
Drive Database}
\usage{
marawi_get_marawi(
  dataset = c("metadata", "demographics", "movements", "culture", "housing",
    "facilities", "establishments", "modetranspo", "conflict", "publicschools",
    "privateschools", "health")
)

marawi_get_marawi_all(tabular = TRUE)
}
\arguments{
\item{dataset}{A string value from either \emph{"metadata"}, \emph{"demographics"},
\emph{"movements"}, \emph{"culture"}, \emph{"housing"}, \emph{"facilities"},
\emph{"establishments"}, \emph{"modetranspo"}, \emph{"conflict"}, \emph{"publicschools"},
\emph{"privateschools"}, and \emph{"health"} specifying the masterlist data to
retrieve. If not specified, will default to \emph{"metadata"}.}

\item{tabular}{Logical. Should output be flattened into a single data.frame?
Default is TRUE. If FALSE, output is a list of data.frames for each
sheet in the masterlist Excel file.}
}
\value{
A tibble or a list of tibbles of the specified masterlist dataset
available in the Lanao del Sur directory of the Open Marawi Database.
}
\description{
Get Marawi Barangay level data masterlist from the Open Marawi Google
Drive Database
}
\examples{
## Retrieve the metadata dataset of the Lanao del Sur directory in the Open
## Marawi database
marawi_get_marawi(dataset = "metadata")
\dontrun{marawi_get_marawi_all()}

}
\author{
Ernest Guevarra
}
