% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_ls.R
\name{marawi_ls}
\alias{marawi_ls}
\alias{marawi_ls_armm}
\alias{marawi_ls_lanao}
\title{List files and folders within the Open Marawi Google Drive database}
\usage{
marawi_ls(id = "1bph1LBRpxwydAvjuggyjmhmq6HhyteY8")

marawi_ls_armm(drive_list)

marawi_ls_lanao(drive_list)
}
\arguments{
\item{id}{Character vector of the Open Marawi \strong{Google Drive} id. This is
currently set to \emph{1bph1LBRpxwydAvjuggyjmhmq6HhyteY8} which is the id of
the Open Marawi \strong{Google Drive}. Change this if Open Marawi
\strong{Google Drive} is moved.}

\item{drive_list}{A dribble from which to search for required Google Drive
directory.}
}
\value{
A dribble of names and ids of files and folders within the
specified Google Drive
}
\description{
A convenience wrapper to \link{googledrive} functions set to specifically access
the Open Marawi Google Drive database and list files and folders within.
}
\examples{
marawi_ls()
marawi_ls_armm(drive_list = marawi_ls())
marawi_ls_lanao(drive_list = marawi_ls())

}
\author{
Ernest Guevarra
}
