% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_maps_get.R
\name{marawi_maps_get}
\alias{marawi_maps_get}
\title{Get maps data from Open Marawi Google Drive}
\usage{
marawi_maps_get(dsn, layer = NULL)
}
\arguments{
\item{dsn}{Data source name of map to retrieve. This should be one of the
data source names listed in \code{marawi_maps_list}}

\item{layer}{Layer name of map to retrieve. If NULL (default), all layers
available within \code{dsn} will be retrieved.}
}
\value{
An sf object of specified map
}
\description{
Get maps data from Open Marawi Google Drive
}
\examples{
##marawi_maps_get(dsn = "Hazards/Faultlines")

}
