% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marawi_maps_list}
\alias{marawi_maps_list}
\title{List of available map files in the Open Marawi Google Drive}
\format{
A data frame with 4 columns and 391 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{dsn} \tab Data source name \cr
   \emph{layer} \tab Layer name \cr
   \emph{name} \tab Filename \cr
   \emph{id} \tab Google Drive identifier \cr
}
}
\usage{
marawi_maps_list
}
\description{
List of available map files in the Open Marawi Google Drive
}
\examples{
marawi_maps_list


}
\keyword{datasets}
