% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_maps_ls.R
\name{marawi_maps_ls}
\alias{marawi_maps_ls}
\title{List Marawi maps}
\usage{
marawi_maps_ls(id = "1Fj5sK2tKo0_v8EnEbN8m1pMfaSFADmYL")
}
\arguments{
\item{id}{Character value of the Open Marawi \strong{Google Drive} id for the
Marawi geospatial directory. This is currently set to
\emph{1Fj5sK2tKo0_v8EnEbN8m1pMfaSFADmYL} which is the id of the current id
of the Open Marawi \strong{Google Drive} for the Marawi geospatial directory.
Change this if Open Marawi \strong{Google Drive} is moved.}
}
\value{
A dribble of all the geospatial files for Marawi found in the
Open Marawi \strong{Google Drive}.
}
\description{
List Marawi maps
}
\examples{
\dontrun{
  marawi_maps_ls()
}

}
