% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_drive_id}
\alias{get_drive_id}
\title{Get Google Drive identifier given a search pattern}
\usage{
get_drive_id(pattern, drive_list)
}
\arguments{
\item{pattern}{A pattern to search for}

\item{drive_list}{A dribble (usually produced from one of the \code{marawi_ls()}
functions) from which to search the \code{pattern} from.}
}
\value{
A Google Drive identifier
}
\description{
Get Google Drive identifier given a search pattern
}
\examples{
get_drive_id("ARMM", drive_list = marawi_ls())

}
