% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmarawi.R
\docType{package}
\name{openmarawi}
\alias{openmarawi}
\alias{openmarawi-package}
\title{An Interface to the Open Marawi Database}
\description{
The citizens of Marawi have a right to the data and maps about their home
city. When problems are complex, helping people find useful maps (access)
can aid them both in finding themselves in the map (understanding) and making
the map by themselves (ownership). Open data and useful maps can help empower
citizens in mapmaking, placemaking, and decision-making because it can help
citizens and interested parties in understanding the issues spatially. It is
practical in deliberating, deciding, and delivering the rehabilitation of
Marawi City.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://panukatan.io/openmarawi/}
  \item \url{https://github.com/panukatan/openmarawi}
  \item Report bugs at \url{https://github.com/panukatan/openmarawi/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
