% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_get.R
\name{marawi_get_armm}
\alias{marawi_get_armm}
\title{Get ARMM Barangay level data masterlist from the Open Marawi Google Drive
Database}
\usage{
marawi_get_armm(
  id = marawi_ls_armm()$id[stringi::stri_detect(marawi_ls_armm()$name, fixed =
    "Masterlist")],
  dataset = c("metadata", "demographics", "facilities", "conflict")
)
}
\arguments{
\item{id}{Character value for the unique identifier for the ARMM Open Marawi
\strong{Google Drive} Masterlist file in \strong{.xlsx} format. Currently set
to the identifier specified when \code{marawi_ls_armm()} is called. Change this
if Open Marawi \strong{Google Drive} is moved or if name of masterlist file
changes.}

\item{dataset}{A string value from either \code{metadata}, \code{demographics},
\code{facilities}, or \code{conflict} specifying the masterlist data to retrieve. If
not specified, will default to \code{metadata}}
}
\value{
A tibble of the specified masterlist dataset available in the
ARMM directory of the Open Marawi Database.
}
\description{
Get ARMM Barangay level data masterlist from the Open Marawi Google Drive
Database
}
\examples{
## Retrieve the metadata dataset of the ARMM directory in the Open Marawi
## database
marawi_get_armm(dataset = "metadata")

}
\author{
Ernest Guevarra
}
