% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marawi_get.R
\name{marawi_ls}
\alias{marawi_ls}
\title{List files and folders within the Open Marawi Google Drive database}
\usage{
marawi_ls(id = "1bph1LBRpxwydAvjuggyjmhmq6HhyteY8")
}
\arguments{
\item{id}{Character vector of the Open Marawi \strong{Google Drive} id. This is
currently set to \emph{1bph1LBRpxwydAvjuggyjmhmq6HhyteY8}}
}
\value{
A tibble of names and ids of files and folders within the the
Open Marawi Google Drive
}
\description{
A convenience wrapper to \link{googledrive} functions set to specifically access
the Open Marawi Google Drive database and list files and folders within.
}
\examples{
marawi_ls()

}
\author{
Ernest Guevarra
}
